♻️ Extract main loop to new module
New InteractiveMode module now house the main loop.

Also, rename `CommandProcessor.process_command` to `call`.
