+++ lib/mix_test_interactive/command_processor.ex
@@ -3,14 +3,14 @@ defmodule MixTestInteractive.CommandProcessor do
 
   alias MixTestInteractive.Config
 
-  @spec process_command(String.t() | :eof, Config.t()) :: {:ok, Config.t()} | :unknown | :quit
+  @spec call(String.t() | :eof, Config.t()) :: {:ok, Config.t()} | :unknown | :quit
 
-  def process_command(:eof, _config), do: :quit
+  def call(:eof, _config), do: :quit
 
-  def process_command(command, config) when is_binary(command) do
+  def call(command, config) when is_binary(command) do
     command
     |> String.trim()
-    |> do_process_command(config)
+    |> process_command(config)
   end
 
   def usage do
@@ -21,7 +21,7 @@ defmodule MixTestInteractive.CommandProcessor do
     """
   end
 
-  defp do_process_command("q", _config), do: :quit
-  defp do_process_command("", config), do: {:ok, config}
-  defp do_process_command(_unknown_command, _config), do: :unknown
+  defp process_command("q", _config), do: :quit
+  defp process_command("", config), do: {:ok, config}
+  defp process_command(_unknown_command, _config), do: :unknown
 end
